Move Ouinet binaries and config into different dir
Use CenoBrowser/Ouinet directory as root for Ouinet client binary and
move ouinet-client.conf and ssl-inj-cert.pem into repo directory. This
is more consistent with TorBrowser's setup. Also do not hard-code the
location of the ssl-inj-cert.pem, pass it as a parameter to the client
executable when Ouinet is started.