Ceno Patch 016: Update installer with Ceno branding
Replace occurrences of Mozilla and Firefox with eQualitie and Ceno in all installer related files.
Change the publisher and trademark of the executables included in the package.
Make additional changes needed for the installer scripts to produce a system-install.
Include Ceno specific files in package-manifest.
Re-add to package-manifest private browsing assets removed by the Tor base-browser patchset.